+++ diesel_compile_tests/tests/compile-fail/returning_cannot_be_called_twice.rs
@@ -9,12 +9,9 @@ table! {
     }
 }
 
-pub struct NewUser(String);
-
-impl_Insertable! {
-    (users)
-    pub struct NewUser(#[column_name(name)] String,);
-}
+#[derive(Insertable)]
+#[table_name = "users"]
+pub struct NewUser(#[column_name = "name"] String);
 
 fn main() {
     use self::users::dsl::*;