+++ diesel_compile_tests/tests/compile-fail/returning_clause_requires_selectable_expression.rs
@@ -10,12 +10,9 @@ table! {
     }
 }
 
-pub struct NewUser(String);
-
-impl_Insertable! {
-    (users)
-    pub struct NewUser(#[column_name(name)] String,);
-}
+#[derive(Insertable)]
+#[table_name = "users"]
+pub struct NewUser(#[column_name = "name"] String);
 
 table! {
     non_users {